requirements.txt
python-dotenv
websocket-client
requests
browser-cookie3
pandas
validators
colorama
emoji
git+https://github.com/rdavydov/Twitch-Channel-Points-Miner-v2.git
